requirements.txt
torch
pytest
pyyaml
line_profiler
einops
kappamodules
matplotlib
